requirements.txt
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==10.0.1
scikit_learn==1.3.1
torch==1.10.2+cu113
torchvision==0.11.3+cu113
